+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 ## Unreleased
+### Fixed
+* Allow usage of `use_camel_case` in client options.
 
 ## 3.0.0